+++ packages/luciq_flutter/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased](https://github.com/luciqai/luciq-flutter-sdk/compare/v18.0.0...dev)
+
+### Added
+
+-  Added `InstabugWidget`, a wrapper used to wrap the main application to provide out of the box automatic Crash Reporting support. ([#3](https://github.com/luciqai/luciq-flutter-sdk/pull/3))
+
 ## [18.0.0](https://github.com/luciqai/luciq-flutter-sdk/compare/v18.0.0...dev) (September 24, 2025)
 
 - SDK rebranded from Instabug to Luciq.